pyproject.toml
[project]
name = "etl-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "db-dtypes>=1.5.0",
    "google-api-python-client>=2.188.0",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-storage>=3.8.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "allure-pytest>=2.15.3",
    "black>=25.11.0",
    "pytest>=8.4.2",
    "pytest-html>=4.1.1",
    "ruff>=0.14.13",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

